# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "DayPoint": {
        "properties": {
          "blocked": {
            "default": 0,
            "title": "Blocked",
            "type": "integer"
          },
          "completed": {
            "default": 0,
            "title": "Completed",
            "type": "integer"
          },
          "cost_usd": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "0",
            "title": "Cost Usd"
          },
          "credits": {
            "default": "0",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Credits",
            "type": "string"
          },
          "day": {
            "title": "Day",
            "type": "string"
          },
          "findings": {
            "default": 0,
            "title": "Findings",
            "type": "integer"
          },
          "reviews": {
            "default": 0,
            "title": "Reviews",
            "type": "integer"
          }
        },
        "required": [
          "day"
        ],
        "title": "DayPoint",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Period": {
        "properties": {
          "days": {
            "title": "Days",
            "type": "integer"
          },
          "since": {
            "format": "date-time",
            "title": "Since",
            "type": "string"
          },
          "until": {
            "format": "date-time",
            "title": "Until",
            "type": "string"
          }
        },
        "required": [
          "since",
          "until",
          "days"
        ],
        "title": "Period",
        "type": "object"
      },
...
      "Usage": {
        "properties": {
          "avg_cost_per_review": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Cost Per Review"
          },
          "avg_tokens_per_review": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Tokens Per Review"
          },
          "balance": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Balance",
            "type": "string"
          },
          "cached_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cached Tokens"
          },
          "chat_credits": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Chat Credits",
            "type": "string"
          },
          "chat_replies": {
            "title": "Chat Replies",
            "type": "integer"
          },
          "completed": {
            "title": "Completed",
            "type": "integer"
          },
          "cost_usd": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd"
          },
          "credits_refunded": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Credits Refunded",
            "type": "string"
          },
          "credits_spent": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Credits Spent",
            "type": "string"
          },
          "daily": {
            "items": {
              "$ref": "#/components/schemas/DayPoint"
            },
            "title": "Daily",
            "type": "array"
          },
          "failed": {
            "title": "Failed",
            "type": "integer"
          },
          "input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "ledger": {
            "items": {
              "$ref": "#/components/schemas/UsageLedgerEntry"
            },
            "title": "Ledger",
            "type": "array"
          },
          "no_credits": {
            "title": "No Credits",
            "type": "integer"
          },
          "output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          },
          "period": {
            "$ref": "#/components/schemas/Period"
          },
          "rate_limited": {
            "title": "Rate Limited",
            "type": "integer"
          },
          "review_events": {
            "title": "Review Events",
            "type": "integer"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/UsageReview"
            },
            "title": "Reviews",
            "type": "array"
          },
          "skipped": {
            "title": "Skipped",
            "type": "integer"
          }
        },
        "required": [
          "period",
          "balance",
          "review_events",
          "completed",
          "rate_limited",
          "no_credits",
          "failed",
          "skipped",
          "credits_spent",
          "credits_refunded",
          "chat_replies",
          "chat_credits",
          "daily",
          "ledger",
          "reviews"
        ],
        "title": "Usage",
        "type": "object"
      },
      "UsageLedgerEntry": {
        "properties": {
          "balance_after": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Balance After",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "delta": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Delta",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "ref_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref Id"
          },
          "ref_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref Type"
          }
        },
        "required": [
          "id",
          "delta",
          "reason",
          "ref_type",
          "ref_id",
          "balance_after",
          "created_at"
        ],
        "title": "UsageLedgerEntry",
        "type": "object"
      },
      "UsageReview": {
        "properties": {
          "cached_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cached Tokens"
          },
          "cost_usd": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "credits_charged": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Credits Charged",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          },
          "pr_number": {
            "title": "Pr Number",
            "type": "integer"
          },
          "repo_full_name": {
            "title": "Repo Full Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "repo_full_name",
          "pr_number",
          "status",
          "credits_charged",
          "created_at"
        ],
        "title": "UsageReview",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "HootPR API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/orgs/{org_slug}/analytics/usage": {
      "get": {
        "operationId": "usage_api_orgs__org_slug__analytics_usage_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_slug",
            "required": true,
            "schema": {
              "title": "Org Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 365,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "repo_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Repo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Usage"
      }
    },
...
    "/api/orgs/{org_slug}/export/{kind}": {
      "get": {
        "operationId": "export_api_orgs__org_slug__export__kind__get",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "reviews",
                "findings",
                "usage",
                "audit"
              ],
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "org_slug",
            "required": true,
            "schema": {
              "title": "Org Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "csv",
              "enum": [
                "csv",
                "json"
              ],
              "title": "Format",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 365,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export"
      }
    },
...
  }
}